Update UsingMetadata.md
On line #3 I think you should consider changing "it" to "the object". The word "it" refers to the object meta data, while the sentence is meant to convey that the object metadata can be set at the time the object is uploaded. 

Thank you for taking time to review my proposed change.