Update Welcome.md
On line #70 the sentence about configuring buckets so that are created in a specific region doesn't really make sense. Buckets are created in a region, they are not configured to be created in a region. It also doesn't make sense to link to the article about Accessing buckets.  The link should be to the article about Creating buckets. 
Thank you for taking the time to review my pull request.