Update versioning-workflows.md
I added a suggestion for line #9. Versioning doesn't actually prevent an object from being overwritten or deleted, it just gives you the ability to restore the object, if necessary, to a previous version. Your application will likely show the object as deleted or overwritten until the object is restored.  
 Actually preventing an object from being overwritten or deleted would involve using an Object Lock - https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-lock.html

Thank you for taking the time to review my pull request.