Update replication-add-config.md
S3 V2 replication schema now has support for delete markers. Suggesting a change to the documentation to make it more clear as it suggests that V2 does not replicate delete markers at all, however it is only rules with tag filters that must be disabled. https://aws.amazon.com/about-aws/whats-new/2020/11/amazon-s3-replication-adds-support-for-replicating-delete-markers/