Make it clear that expirations are not charged
Lifecycle transitions are charged, but expirations are not. Since they are configured in the same place it's not easy to know that expirations are in fact free. The only charges are storage in some situations. This change makes it clear that the charges described further down are only about storage.