Add <!DOCTYPE html> to example index.html
Chrome 92.0.4515.107 was downloading the document instead of rendering it as a web page even when Content-Type on the S3 object was configured to "text/html" and S3 bucket configured to for hosting a static website.   Adding the <!DOCYTYPE html> line caused the browser to become happy to render it.

Proposing this change in example documentation to help someone else avoid this confusion.